fixed: baja de productos

+++ login/ABM/borCategoria.php
@@ -1,46 +1,28 @@
 <?php
 
-if(mysqli_connect('localhost','root','','login')){
+if (mysqli_connect('localhost', 'root', '', 'login')) {
 	//conexion a la base de datos con 4 parametros (servidor, usuarioServidor, contraseña usuario y nombre de la base de datos)
 
-	
-	
-	if(isset($_GET['categoria']) and isset($_GET['nombre'])){
+	if (isset($_GET['categoria']) and isset($_GET['nombre'])) {
 		$codigo = $_GET['categoria'];
 		$nombre = $_GET['nombre'];
-		
-		$con = mysqli_connect('localhost','root','','login');
+
+		$con = mysqli_connect('localhost', 'root', '', 'login');
 		//guardo los datos de conexion
-		
-		$consulta = "DELETE FROM categorias WHERE idCategoria='$codigo'";
-		//guardo la consulta que quiero hacerle a la base de datos
-		
-		if($resultado = mysqli_query($con, $consulta)){
-			//guardo el resultado de la consulta de base de datos
+
+		$consultaProductos = "DELETE FROM productos WHERE categoriaProducto='$codigo'";
+		if ($resultadoProductos = mysqli_query($con, $consultaProductos)) {
+			$consulta = "DELETE FROM categorias WHERE idCategoria='$codigo'";
+			if ($resultado = mysqli_query($con, $consulta)) {
 				print "<h1>La categoria $nombre fue Eliminada :C </h1>";
 				print "<p><a href='index.php'>Inicio</a></p>";
-				
-			
+			}
 		}
-		
-		
-			
-		
-	}else{
+
+	} else {
 		print "<h1>Algo se rompio</h1>";
 	}
-		
-		
-		
-		
-		
-	
-	
-	
-	
+
 }
-
-
-
 
 ?>
